refactor: make different backends for storage
Storage is now handled similarly to how we handle other parts of our
infrastructure where the provider is defined by the runtime environment:

- a thin dispatch layer is responsible for understand the configuration
  and call the right provider
- the provider adheres to a shared behaviour.
